Commit 20131119

+++ main/ZimbraWebClient/WebRoot/js/ajax/dwt/widgets/DwtMenu.js
@@ -1,10 +1,10 @@
 /*
  * ***** BEGIN LICENSE BLOCK *****
  * Zimbra Collaboration Suite Web Client
- * Copyright (C) 2005, 2006, 2007, 2008, 2009, 2010, 2011, 2012 VMware, Inc.
+ * Copyright (C) 2005, 2006, 2007, 2008, 2009, 2010, 2011, 2012, 2013 Zimbra Software, LLC.
  * 
  * The contents of this file are subject to the Zimbra Public License
- * Version 1.3 ("License"); you may not use this file except in
+ * Version 1.4 ("License"); you may not use this file except in
  * compliance with the License.  You may obtain a copy of the License at
  * http://www.zimbra.com/license.
  * 
@@ -133,10 +133,8 @@ DwtMenu = function(params) {
 	this.__currentItem = null;
 	this.__preventMenuFocus = false;
 
-	// Default menu tab group. Note that we disable application handling of
-	// keyboard shortcuts, since we don't want the view underneath reacting to
-	// keystrokes in the menu.
-	this._tabGroup = new DwtTabGroup(this.toString(), true);
+	// Default menu tab group.
+	this._tabGroup = new DwtTabGroup(this.toString());
 	this._tabGroup.addMember(this);
 	this._created = true;
 
@@ -917,7 +915,7 @@ function(child) {
 			this._table.rows[0].deleteCell(Dwt.getCellIndex(cell));
 		} else {
 			var el = child.getHtmlElement();
-			if (el && el.parentNode.parentNode.rowIndex > -1)// Make sure that the element exists in the table
+			if (el && el.parentNode && el.parentNode.parentNode.rowIndex > -1)// Make sure that the element exists in the table
 				this._table.deleteRow(el.parentNode.parentNode.rowIndex);
 		}
 	}
